% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{add_environmental_noise}
\alias{add_environmental_noise}
\title{Add environmental variation to get trait values.}
\usage{
add_environmental_noise(genetic_value, Ve)
}
\arguments{
\item{genetic_value}{Genetic values for individuals.}

\item{Ve}{Desired environmental variance.}
}
\value{
Vector of trait values.
}
\description{
Add environmental variation to get trait values.
}
