% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{balanced_cross}
\alias{balanced_cross}
\title{Randomly mate individuals from a population while assigning the number
of offspring in a balanced way across parents and sexes}
\usage{
balanced_cross(females, males, n_crosses, n_progeny, simparam)
}
\arguments{
\item{females}{Population object of mothers.}

\item{males}{Population object of fathers.}

\item{n_crosses}{Number of crosses.}

\item{n_progeny}{Number of offspring per cross.}

\item{simparam}{Simulation parameters object.}
}
\value{
Population object of offspring.
}
\description{
In this crossing scheme, males and females are assigned as fathers and mothers
with equal contributions within sex. The contributions between sexes will be
different if the number of males and females used differs.
}
