% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{c.population}
\alias{c.population}
\title{Concatenate two populations}
\usage{
\method{c}{population}(pop1, pop2)
}
\arguments{
\item{pop1}{First population}

\item{pop2}{Second population}
}
\value{
Combined population
}
\description{
Concatenate two populations
}
