% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{calculate_genetic_values}
\alias{calculate_genetic_values}
\title{Get genetic values from genotypes and locus-level coefficients.}
\usage{
calculate_genetic_values(geno, a, d)
}
\arguments{
\item{geno}{Matrix of individual x locus genotypes coded 0, 1, 2.}

\item{a}{Vector of additive genetic coefficients.}

\item{d}{Vector of dominance coefficients.}
}
\value{
Vector of genetic values.
}
\description{
Get genetic values from genotypes and locus-level coefficients.
}
