% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{draw_founder_genotypes_freq}
\alias{draw_founder_genotypes_freq}
\title{Sample founder genotypes with given frequencies.}
\usage{
draw_founder_genotypes_freq(n_ind, freq)
}
\arguments{
\item{n_ind}{Number of individuals.}

\item{freq}{Vector of base population frequencies.}
}
\value{
Matrix of individual x locus genotypes coded 0, 1, 2
}
\description{
Sample founder genotypes with given frequencies.
}
