% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{get_females}
\alias{get_females}
\title{Extract females from a population.}
\usage{
get_females(pop)
}
\arguments{
\item{pop}{A population object.}
}
\value{
Population of females.
}
\description{
Extract females from a population.
}
