% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{get_males}
\alias{get_males}
\title{Extract males from a population.}
\usage{
get_males(pop)
}
\arguments{
\item{pop}{A population object.}
}
\value{
Population of males.
}
\description{
Extract males from a population.
}
