% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{make_cross}
\alias{make_cross}
\title{Make crosses according to a cross plan}
\usage{
make_cross(pop, cross_plan, n_progeny, simparam)
}
\arguments{
\item{pop}{Population to take individuals from.}

\item{cross_plan}{Matrix of individual IDs, where the first column is
mothers and the second fathers.}

\item{n_progeny}{Number of progenies per cross.}

\item{simparam}{Simulation parameters.}
}
\value{
Population object of offspring.
}
\description{
Crosses are made based on a matrix of parents. This function is used
internally by other crossing functions, and can be used if you want
more control over the cross plan.
}
