% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_trait}
\alias{make_trait}
\title{Create a quantitative trait based on founder genotypes.}
\usage{
make_trait(
  founder_geno,
  n_qtl,
  distribution,
  parameters,
  Vg = 1,
  trait_mean = 0
)
}
\arguments{
\item{founder_geno}{Matrix of individual x locus founder genotypes coded 0, 1, 2.}

\item{n_qtl}{Number of causative variants desired.}

\item{distribution}{Distribution for additive and dominance coefficients; only
"gaussian" is implemented.}

\item{parameters}{List of parameters with named components: mean_a and
var_a are the mean and variance for he additive genetic coefficients;
mean_dd and var_dd are the, optional, mean and variance for dominance degrees.}

\item{Vg}{Desired genetic variance.}
}
\value{
A list describing the trait, with indices to the loci, and additive
genetic coefficients and dominance coefficients for each locus.
}
\description{
Create a quantitative trait based on founder genotypes.
}
