% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_trait_manual}
\alias{make_trait_manual}
\title{Create a trait based on given effect sizes}
\usage{
make_trait_manual(founder_geno, loci_ix, a, d, Vg = NULL, trait_mean = NULL)
}
\arguments{
\item{founder_geno}{Matrix of individual x locus founder genotypes coded 0, 1, 2.}

\item{loci_ix}{Indices of causative variants in genotype matrix.}

\item{a}{Vector of additive genetic coefficients.}

\item{d}{Vector of dominance coefficients.}

\item{Vg}{Desired genetic variance; NULL for no adjustment.}

\item{trait_mean}{Desired trait mean; NULL for no adjustment.}
}
\value{
A list describing the trait, with indices to the loci, and additive
genetic coefficients and dominance coefficients for each locus.
}
\description{
This function is used internally by make_trait, and can be used if you
want more control of genetic architecture.
}
\details{
Polymorphic loci from the founder genotypes are randomly chosen to be
causative and assigned additive genetic coefficients and optionally
dominance coefficients based on a probability distribution.

The resulting genetic variance and trait mean are adjusted to
desired numbers.
}
