% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{meanG}
\alias{meanG}
\title{Get mean genetic value from population object.}
\usage{
meanG(pop)
}
\arguments{
\item{pop}{Population object.}
}
\value{
Mean.
}
\description{
Get mean genetic value from population object.
}
