% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutate_new_locus}
\alias{mutate_new_locus}
\title{Mutate the genotype matrix of a population by introducing new mutations}
\usage{
mutate_new_locus(pop, simparam, mutation_rate, var_a, trait = 1)
}
\arguments{
\item{pop}{Population object.}

\item{simparam}{Simulation parameters object.}

\item{mutation_rate}{Mutation rate per individual and generation.}

\item{var_a}{Variance of the additive genetic coefficient.}

\item{trait}{Trait number to be mutated. For now, always 1.}
}
\value{
List containing population object (first component) and simparam
object (second component) with mutated alleles changed and trait information
updated.
}
\description{
This function introduces mutations to a population object. Mutations are assumed
to always introduce a new locus, with a causative allele in one copy. There
is no backward mutation. The mutation's additive effect is drawn from a normal
distribution with given variance. The dominance coefficient is always zero.
}
