% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutate_per_locus}
\alias{mutate_per_locus}
\title{Mutate the genotype matrix of a population by a per-locus mutation rate.}
\usage{
mutate_per_locus(pop, mutation_rate)
}
\arguments{
\item{pop}{Population object.}

\item{mutation_rate}{Mutation rate per locus and genome copy.}
}
\value{
Population object with mutated alleles changed.
}
\description{
This function introduces mutations to a population object. Mutations are
assumed to happen between the two possible alleles with the same probabilty
for forward and backward mutation. The mutation rate is per locus.
}
