% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{new_population}
\alias{new_population}
\title{Create a population object}
\usage{
new_population(geno, simparam)
}
\arguments{
\item{geno}{Matrix of genotypes.}

\item{simparam}{Simulation parameters object.}
}
\value{
A population object, which contains the following components:
 \itemize{
 \item{"id"}{ with the individuals' identifiers}
 \item{"sex"}{ with sexes coded F and M}
 \item{"geno}{ with the genotype matrix}
 \item{"gv"}{ with genetic values}
 \item{"pheno"}{ with trait values}
}
}
\description{
Create a population object
}
