% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{pick_loci}
\alias{pick_loci}
\title{Pick candidate loci to be causative variants from a genotype matrix.}
\usage{
pick_loci(geno, n_qtl)
}
\arguments{
\item{geno}{Matrix of individual x locus genotypes coded 0, 1, 2.}

\item{n_qtl}{Number of causative variants desired.}
}
\value{
Indices to selected loci.
}
\description{
Pick candidate loci to be causative variants from a genotype matrix.
}
