% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{print.population}
\alias{print.population}
\title{Print a nice summary of a population object.}
\usage{
\method{print}{population}(pop)
}
\arguments{
\item{pop}{}
}
\value{
Nothing.
}
\description{
Print a nice summary of a population object.
}
