% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{pull_freq}
\alias{pull_freq}
\title{Get frequencies of all variants from population object.}
\usage{
pull_freq(pop)
}
\arguments{
\item{pop}{Population object.}
}
\value{
Allele frequencies
}
\description{
Get frequencies of all variants from population object.
}
