% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{pull_nonqtl_freq}
\alias{pull_nonqtl_freq}
\title{Get frequencies of non-causative variants from population object}
\usage{
pull_nonqtl_freq(pop, simparam)
}
\arguments{
\item{pop}{Population object.}

\item{simparam}{Simulation parameters object.}
}
\value{
Allele frequencies
}
\description{
Get frequencies of non-causative variants from population object
}
