% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{pull_qtl_freq}
\alias{pull_qtl_freq}
\title{Get frequencies of causative variants from population object}
\usage{
pull_qtl_freq(pop, trait = 1, simparam)
}
\arguments{
\item{pop}{Population object.}

\item{trait}{Trait number.}

\item{simparam}{Simulation parameters object.}
}
\value{
Allele frequencies
}
\description{
Get frequencies of causative variants from population object
}
