% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{pull_qtl_geno}
\alias{pull_qtl_geno}
\title{Get genotypes at causative variants from population object}
\usage{
pull_qtl_geno(pop, trait = 1, simparam)
}
\arguments{
\item{pop}{Population object.}

\item{trait}{Trait number.}

\item{simparam}{Simulation parameters object.}
}
\value{
Genotype matrix coded 0, 1, 2.
}
\description{
Get genotypes at causative variants from population object
}
