% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{rand_cross}
\alias{rand_cross}
\title{Assign random crosses where each parent, within sex, is equally likely to be
chosen each time.}
\usage{
rand_cross(females, males, n_crosses, n_progeny, simparam)
}
\arguments{
\item{females}{Population of mothers.}

\item{males}{Populations of fathers.}

\item{n_crosses}{Number of crosses.}

\item{n_progeny}{Number of offspring per cross.}

\item{simparam}{Simulation parameters object.}
}
\value{
Population object of offspring.
}
\description{
Assign random crosses where each parent, within sex, is equally likely to be
chosen each time.
}
