% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{select_cross_fitness}
\alias{select_cross_fitness}
\title{Cross individuals in a population based on sampling weighted by fitness}
\usage{
select_cross_fitness(
  pop,
  n_crosses,
  n_progeny,
  fitness_function,
  simparam,
  trait = 1
)
}
\arguments{
\item{pop}{Population of parents}

\item{n_crosses}{Number of crosses.}

\item{n_progeny}{Number of offspring per cross.}

\item{fitness_function}{A function that takes trait values and return
fitness values.}

\item{simparam}{Simulation parameters object.}

\item{trait}{Trait number of the trait to select on.}
}
\value{
A population object of selected individuals.
}
\description{
Cross individuals in a population based on sampling weighted by fitness
}
