% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{select_culling}
\alias{select_culling}
\title{Select individuals from a population that cross a given threshold}
\usage{
select_culling(pop, n_ind, threshold, method = "pheno", trait = 1)
}
\arguments{
\item{pop}{Population to select from.}

\item{n_ind}{Number of individuals to select.}

\item{threshold}{Phenotypic threshold for inclusion.}

\item{method}{Selection method. Only "pheno" for trait values is implemented.}

\item{trait}{Trait number of the trait to select on.}
}
\value{
A population object of selected individuals.
}
\description{
This function selects a number of individuals by independent culling,
only allowing individuals whose trait values are above a given threshold.
When there are more individuals available than given, eligible individuals
are sampled randomly without replacement.
}
