% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{select_ind}
\alias{select_ind}
\title{Select top individuals from a population based on trait}
\usage{
select_ind(pop, n_ind, method = "pheno", trait = 1)
}
\arguments{
\item{pop}{Population to select from.}

\item{n_ind}{Number of individuals to select.}

\item{method}{Selection method. "pheno" for trait values. "rand" for random.}

\item{trait}{Trait number of the trait to select on.}
}
\value{
A population object of selected individuals.
}
\description{
This function selects a number of individuals based on
either the ranking of their trait values or at random.
High trait values are selected.
}
