% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{select_ind_fitness}
\alias{select_ind_fitness}
\title{Select individuals from a population based on sampling weighted by
a fitness function.}
\usage{
select_ind_fitness(pop, n_ind, fitness_function, trait = 1)
}
\arguments{
\item{pop}{Population to select from.}

\item{n_ind}{Number of individuals to select.}

\item{fitness_function}{A function that takes trait values and return
fitness values.}

\item{trait}{Trait number of the trait to select on.}
}
\value{
A population object of selected individuals.
}
\description{
Select individuals from a population based on sampling weighted by
a fitness function.
}
