% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{SimParam}
\alias{SimParam}
\title{Simulation parameters object}
\description{
Contains data on traits, environmental variance and settings for a
simsimqt simulation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SimParam-new}{\code{SimParam$new()}}
\item \href{#method-SimParam-clone}{\code{SimParam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimParam-new"></a>}}
\if{latex}{\out{\hypertarget{method-SimParam-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimParam$new(traits = traits, Ve = Ve, use_sexes = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SimParam-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SimParam-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimParam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
