% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{[.population}
\alias{[.population}
\title{Hard bracket indexing on a population object takes a subset of individuals.}
\usage{
\method{[}{population}(pop, ids)
}
\arguments{
\item{pop}{A population object}

\item{ids}{Individual IDs to select.}
}
\value{
A population object containing the subset.
}
\description{
Hard bracket indexing on a population object takes a subset of individuals.
}
