% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{validate_population}
\alias{validate_population}
\title{Check the validity of a population object.}
\usage{
validate_population(pop)
}
\arguments{
\item{pop}{A population object.}
}
\value{
The object, unchanged.
}
\description{
Check the validity of a population object.
}
