% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{varG}
\alias{varG}
\title{Get genetic variance from population object}
\usage{
varG(pop)
}
\arguments{
\item{pop}{Population object.}
}
\value{
Variance.
}
\description{
Get genetic variance from population object
}
