% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{draw_founder_genotypes}
\alias{draw_founder_genotypes}
\title{Sample founder genotypes with frequencies drawn from a random
distribution.}
\usage{
draw_founder_genotypes(n_ind, n_loci, distribution, parameters)
}
\arguments{
\item{n_ind}{Number of individuals.}

\item{n_loci}{Number of loci (note that they are not guaranteed to be polymorphic,
so this number should be bigger than the number of causative variants that
are planned)}

\item{distribution}{Distribution for allele frequencies; only beta is implemented.}

\item{parameters}{List of parameters for the random distribution (e.g., for beta
it would be list(shape1, shape2).}
}
\value{
Matrix of individual x locus genotypes coded 0, 1, 2
}
\description{
Sample founder genotypes with frequencies drawn from a random
distribution.
}
