% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{make_trait_manual}
\alias{make_trait_manual}
\title{Create a trait based on given effect sizes.}
\usage{
make_trait_manual(founder_geno, loci_ix, a, d, Vg = NULL, trait_mean = NULL)
}
\arguments{
\item{founder_geno}{Matrix of individual x locus founder genotypes coded 0, 1, 2.}

\item{loci_ix}{Indices of causative variants in genotype matrix.}

\item{a}{Vector of additive genetic coefficients.}

\item{d}{Vector of dominance coefficients.}

\item{Vg}{Desired genetic variance; NULL for no adjustment.}

\item{trait_mean}{Desired trait mean; NULL for no adjustment.}
}
\value{
A list describing the trait, with indices to the loci, and additive
genetic coefficients and dominance coefficients for each locus.
}
\description{
Create a trait based on given effect sizes.
}
