% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{new_population}
\alias{new_population}
\title{Create a population object}
\usage{
new_population(geno, simparam)
}
\arguments{
\item{geno}{Matrix of genotypes.}

\item{simparam}{Simulation parameters object.}
}
\value{

}
\description{
Create a population object
}
