% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{rand_cross}
\alias{rand_cross}
\title{Assign random crosses where each parent, regardless of sex, is equally likely
to be chosen each time.}
\usage{
rand_cross(pop, n_crosses, n_progeny, simparam)
}
\arguments{
\item{pop}{Population of parents}

\item{n_crosses}{Number of crosses.}

\item{n_progeny}{Number of offspring per cross.}

\item{simparam}{Simulation parameters object.}
}
\value{
Population object of offspring.
}
\description{
Assign random crosses where each parent, regardless of sex, is equally likely
to be chosen each time.
}
