% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_structure.R
\name{select_ind}
\alias{select_ind}
\title{Select top individuals from a population based on trait}
\usage{
select_ind(pop, n_ind, trait = 1)
}
\arguments{
\item{pop}{Population to select from.}

\item{n_ind}{Number of individuals to select.}

\item{trait}{Trait number of the trait to select on.}
}
\value{
A population object of selected individuals.
}
\description{
Select top individuals from a population based on trait
}
