% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{SimParam}
\alias{SimParam}
\title{Create the simulation parameters object. Currently very uninteresting.}
\value{
Simulation parameters object.
}
\description{
Create the simulation parameters object. Currently very uninteresting.

Create the simulation parameters object. Currently very uninteresting.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SimParam$new()}}
\item \href{#method-clone}{\code{SimParam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimParam$new(traits = traits, Ve = Ve, use_sexes = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{traits}}{List of traits.}

\item{\code{Ve}}{Vector of environmental variances if intented to be constant.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SimParam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
